+++ ADMF/internal/functions/Resolve-DomainController.ps1
@@ -38,7 +38,7 @@
 		[PSCredential]
 		$Credential,
 
-		[ValidateSet('PDCEmulator', 'Random')]
+		[ValidateSet('PDCEmulator', 'Site', 'Random')]
 		[string]
 		$Type = (Get-PSFConfigValue -FullName 'ADMF.DCSelectionMode' -Fallback 'PDCEmulator')
 	)
@@ -73,6 +73,27 @@
 				Write-PSFMessage -Level Host -String 'Resolve-DomainController.Resolved' -StringValues $domainController.HostName
 				$script:resolvedDomainController = $domainController.HostName
 				return $domainController.HostName
+			}
+			'Site' {
+				$allDC = Get-ADDomainController @parameters -Filter *
+				$targetDCs = $allDC | Where-Object Site -EQ (Get-PSFConfigValue -FullName 'ADMF.DCSelection.Site')
+				$domain = Get-ADDomain @parameters
+				if ($targetDCs.HostName -contains $domain.PDCEmulator) {
+					Write-PSFMessage -Level Host -String 'Resolve-DomainController.Resolved' -StringValues $domain.PDCEmulator
+					$script:resolvedDomainController = $domain.PDCEmulator
+					return $domain.PDCEmulator
+				}
+				elseif ($targetDCs) {
+					$dcObject = $targetDCs | Get-Random
+					Write-PSFMessage -Level Host -String 'Resolve-DomainController.Resolved' -StringValues $dcObject.HostName
+					$script:resolvedDomainController = $dcObject.HostName
+					return $dcObject.HostName
+				}
+				else {
+					Write-PSFMessage -Level Host -String 'Resolve-DomainController.Resolved' -StringValues $domain.PDCEmulator
+					$script:resolvedDomainController = $domain.PDCEmulator
+					$domain.PDCEmulator
+				}
 			}
 			default {
 				$domain = Get-ADDomain @parameters